refactor: Accurate filenames, nullability.
